+++ ios/Framework/Themes/Default/Portlet/cordova/ScreensCordovaViewController.swift
@@ -17,20 +17,19 @@ import Cordova
 import WebKit
 
 // swiftlint:disable weak_delegate
-public class ScreensCordovaViewController: CDVViewController, UIWebViewDelegate, WKNavigationDelegate {
+public class ScreensCordovaViewController: CDVViewController, WKNavigationDelegate {
 
-	var cdvDelegate: CDVUIWebViewNavigationDelegate?
 	var initialNavigation: WKNavigation?
 	lazy var wkUIDelegate: ScreensWKUIDelegate = ScreensWKUIDelegate(viewController: self)
 	var wkDelegate: WKNavigationDelegate? {
 		return webViewEngine as? WKNavigationDelegate
 	}
 
-	let jsCallHandler: (String, String) -> Void
-	let onPageLoadFinished: (String, Error?) -> Void
+	let jsCallHandler: ScreensWebView.JsCallHandler
+	let onPageLoadFinished: ScreensWebView.OnPageLoadFinished
 
-	public init(jsCallHandler: @escaping (String, String) -> Void,
-		onPageLoadFinished: @escaping (String, Error?) -> Void) {
+	public init(jsCallHandler: @escaping ScreensWebView.JsCallHandler,
+		onPageLoadFinished: @escaping ScreensWebView.OnPageLoadFinished) {
 
 		self.jsCallHandler = jsCallHandler
 		self.onPageLoadFinished = onPageLoadFinished
@@ -43,16 +42,18 @@ public class ScreensCordovaViewController: CDVViewController, UIWebViewDelegate,
 
 	public override func viewDidLoad() {
 		super.viewDidLoad()
-		register(ScreensBridgePlugin(webViewEngine: self.webViewEngine), withPluginName: "screensbridgeplugin")
-		cdvDelegate = CDVUIWebViewNavigationDelegate(enginePlugin: self.webViewEngine as! CDVPlugin)
-
-		if let wkWebView = self.webViewEngine.engineWebView as? WKWebView {
-			wkWebView.uiDelegate = wkUIDelegate
+		guard let webView = self.webViewEngine.engineWebView as? WKWebView else {
+			fatalError("You have to install the WKWebView plugin")
 		}
+
+		register(ScreensBridgePlugin(webViewEngine: self.webViewEngine), withPluginName: "screensbridgeplugin")
+
+		webView.uiDelegate = wkUIDelegate
+		webView.configuration.userContentController.add(
+			WeakMessageHandler(jsCallHandler: jsCallHandler), name: "screensDefault")
 	}
 
 	public func inject(script: InjectableScript, completionHandler: ((Any?, Error?) -> Void)?) {
-
 		self.webViewEngine.evaluateJavaScript(script.content, completionHandler: completionHandler)
 	}
 
@@ -65,36 +66,24 @@ public class ScreensCordovaViewController: CDVViewController, UIWebViewDelegate,
 		initialNavigation = webViewEngine.loadHTMLString(htmlString, baseURL: URL(string: server)!) as? WKNavigation
 	}
 
-	// MARK: UIWebViewDelegate
-
-	public func webView(_ webView: UIWebView,
-	                    shouldStartLoadWith request: URLRequest, navigationType: UIWebViewNavigationType) -> Bool {
-		return cdvDelegate?.webView(webView, shouldStartLoadWith: request, navigationType: navigationType) ?? false
-
-	}
-
-	public func handleJsCall(namespace: String, message: String) {
+	open func handleJsCall(namespace: String, message: String) {
 		jsCallHandler(namespace, message)
-	}
-
-	public func webViewDidStartLoad(_ webView: UIWebView) {
-		cdvDelegate?.webViewDidStartLoad(webView)
-	}
-
-	public func webViewDidFinishLoad(_ webView: UIWebView) {
-		cdvDelegate?.webViewDidFinishLoad(webView)
-		onPageLoadFinished(webView.request?.url?.absoluteString ?? "", nil)
-	}
-
-	public func webView(_ webView: UIWebView, didFailLoadWithError error: Error) {
-		cdvDelegate?.webView(webView, didFailLoadWithError: error)
-		onPageLoadFinished(webView.request?.url?.absoluteString ?? "", error)
 	}
 
 	// MARK: WKNavigationDelegate
 
 	public func webView(_ webView: WKWebView, didStartProvisionalNavigation navigation: WKNavigation!) {
 		wkDelegate?.webView?(webView, didStartProvisionalNavigation: navigation)
+	}
+
+	public func webView(_ webView: WKWebView, didCommit navigation: WKNavigation!) {
+		if initialNavigation == nil || initialNavigation != navigation {
+			let notifyDOMContentLoaded = "document.addEventListener('DOMContentLoaded', function(event) {"
+				+ "window.webkit.messageHandlers.screensDefault.postMessage(['DOMContentLoaded', '']);"
+				+ "});"
+
+			webView.evaluateJavaScript(notifyDOMContentLoaded, completionHandler: nil)
+		}
 	}
 
 	public func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
@@ -126,5 +115,21 @@ public class ScreensCordovaViewController: CDVViewController, UIWebViewDelegate,
 		}
 
 		wkDelegate?.webView?(webView, decidePolicyFor: navigationAction, decisionHandler: decisionHandler)
+	}
+}
+
+public class WeakMessageHandler: NSObject, WKScriptMessageHandler {
+	let jsCallHandler: (String, String) -> Void
+
+	public init(jsCallHandler: @escaping (String, String) -> Void) {
+		self.jsCallHandler = jsCallHandler
+	}
+
+	public func userContentController(_ userContentController: WKUserContentController,
+			didReceive message: WKScriptMessage) {
+
+		guard let body = message.body as? [String] else { return }
+
+		jsCallHandler(body[0], body[1])
 	}
 }